# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/locations": {
      "get": {
        "description": "Para los selects en cascada: ?parentId=<provincia> trae sus localidades.",
        "operationId": "LocationsController_list_api/v1",
        "parameters": [
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "true (por defecto): sólo activos, los de los selectores. false: sólo desactivados. all: todos.",
            "schema": {
              "default": "true",
              "type": "string",
              "enum": [
                "true",
                "false",
                "all"
              ]
            }
          },
          {
            "name": "level",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "COUNTRY",
                "PROVINCE",
                "CITY",
                "NEIGHBORHOOD"
              ]
            }
          },
          {
            "name": "parentId",
            "required": false,
            "in": "query",
            "description": "Sólo las que están directamente dentro de esta ubicación",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LocationDto"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Rol insuficiente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar, filtrando por nivel y por padre",
        "tags": [
          "Master data"
        ]
      },
      "post": {
        "operationId": "LocationsController_create_api/v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLocationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocationDto"
                }
              }
            }
          },
          "400": {
            "description": "Falta el padre (todo menos un país lo necesita)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Rol insuficiente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "Ya existe con ese nombre en el mismo padre",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "El padre no existe o no puede contenerla",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crear (sólo admin)",
        "tags": [
          "Master data"
        ]
      }
    },
    "/api/v1/locations/tree": {
      "get": {
        "operationId": "LocationsController_tree_api/v1",
        "parameters": [
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "true (por defecto): sólo activos, los de los selectores. false: sólo desactivados. all: todos.",
            "schema": {
              "default": "true",
              "type": "string",
              "enum": [
                "true",
                "false",
                "all"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LocationNodeDto"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Rol insuficiente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "El árbol completo, anidado",
        "tags": [
          "Master data"
        ]
      }
    },
    "/api/v1/locations/{id}": {
      "get": {
        "operationId": "LocationsController_findOne_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocationDto"
                }
              }
            }
          },
          "403": {
            "description": "Rol insuficiente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Ver una",
        "tags": [
          "Master data"
        ]
      },
      "patch": {
        "operationId": "LocationsController_rename_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameLocationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocationDto"
                }
              }
            }
          },
          "403": {
            "description": "Rol insuficiente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Renombrar (sólo admin)",
        "tags": [
          "Master data"
        ]
      }
    },
    "/api/v1/locations/{id}/deactivate": {
      "patch": {
        "operationId": "LocationsController_deactivate_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocationDto"
                }
              }
            }
          },
          "403": {
            "description": "Rol insuficiente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Desactivar (sólo admin)",
        "tags": [
          "Master data"
        ]
      }
    },
    "/api/v1/locations/{id}/activate": {
      "patch": {
        "description": "Se permite aunque el padre esté desactivado.",
        "operationId": "LocationsController_activate_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocationDto"
                }
              }
            }
          },
          "403": {
            "description": "Rol insuficiente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reactivar (sólo admin)",
        "tags": [
          "Master data"
        ]
      }
    },
...
  },
  "info": {
    "title": "Adminprop API",
    "description": "API NestJS con arquitectura hexagonal, CQRS y autenticación JWT propia.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Access token devuelto por POST /api/v1/auth/login"
      }
    },
    "schemas": {
      "ApiErrorDetailDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Descripción legible del problema puntual",
            "example": "El email no tiene un formato válido"
          },
          "code": {
            "type": "string",
            "description": "Código de la regla que falló",
            "example": "isEmail"
          },
          "field": {
            "type": "string",
            "description": "Campo que originó el error",
            "example": "email"
          }
        },
        "required": [
          "message"
        ]
      },
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Indica si la operación fue exitosa",
            "example": false
          },
          "statusCode": {
            "type": "number",
            "description": "Status HTTP de la respuesta",
            "example": 400
          },
          "error": {
            "type": "string",
            "description": "Nombre del status HTTP",
            "example": "Bad Request"
          },
          "code": {
            "type": "string",
            "description": "Código de error estable. Ramificá por acá, no por el texto del mensaje.",
            "enum": [
              "INTERNAL_ERROR",
              "BAD_REQUEST",
              "VALIDATION_FAILED",
              "NOT_FOUND",
              "CONFLICT",
              "UNPROCESSABLE_ENTITY",
              "PAYLOAD_TOO_LARGE",
              "UNSUPPORTED_MEDIA_TYPE",
              "METHOD_NOT_ALLOWED",
              "RATE_LIMIT_EXCEEDED",
              "SERVICE_UNAVAILABLE",
              "REQUEST_TIMEOUT",
              "UNAUTHENTICATED",
              "INVALID_CREDENTIALS",
              "TOKEN_MISSING",
              "TOKEN_INVALID",
              "TOKEN_EXPIRED",
              "REFRESH_TOKEN_INVALID",
              "REFRESH_TOKEN_EXPIRED",
              "REFRESH_TOKEN_REUSED",
              "SESSION_REVOKED",
              "EMAIL_ALREADY_REGISTERED",
              "ACCOUNT_INACTIVE",
              "ACCOUNT_LOCKED",
              "WEAK_PASSWORD",
              "CURRENT_PASSWORD_INVALID",
              "PASSWORD_REUSED",
              "FORBIDDEN",
              "INSUFFICIENT_PERMISSIONS",
              "FEATURE_DISABLED",
              "DATABASE_ERROR",
              "UNIQUE_CONSTRAINT_VIOLATION",
              "FOREIGN_KEY_CONSTRAINT_VIOLATION",
              "RECORD_NOT_FOUND",
              "RELATED_RECORDS_EXIST",
              "EXTERNAL_SERVICE_ERROR",
              "EXTERNAL_SERVICE_TIMEOUT",
              "EXTERNAL_SERVICE_UNAVAILABLE"
            ],
            "example": "VALIDATION_FAILED"
          },
          "message": {
            "type": "string",
            "description": "Mensaje legible que describe el error",
            "example": "La validación de los datos enviados falló"
          },
          "errors": {
            "description": "Detalle por campo. Vacío si el error no es de validación.",
            "example": [
              {
                "message": "El email no tiene un formato válido",
                "code": "isEmail",
                "field": "email"
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiErrorDetailDto"
            }
          },
          "correlationId": {
            "type": "string",
            "description": "ID de correlación del request, para rastrearlo en los logs",
            "example": "3f8c1b1e-3a6f-4f2e-9a1c-6d9f2b7c4e11"
          },
          "timestamp": {
            "type": "string",
            "description": "Momento en que se generó el error (ISO 8601)",
            "example": "2026-01-15T10:00:00.000Z"
          },
          "path": {
            "type": "string",
            "description": "Path del request que falló",
            "example": "/api/v1/auth/login"
          }
        },
        "required": [
          "success",
          "statusCode",
          "error",
          "code",
          "message",
          "errors",
          "correlationId",
          "timestamp",
          "path"
        ]
      },
...
      "LocationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "3f8c1b1e-3a6f-4f2e-9a1c-6d9f2b7c4e11"
          },
          "level": {
            "type": "string",
            "enum": [
              "COUNTRY",
              "PROVINCE",
              "CITY",
              "NEIGHBORHOOD"
            ],
            "example": "NEIGHBORHOOD"
          },
          "name": {
            "type": "string",
            "example": "Palermo"
          },
          "parentId": {
            "type": "string",
            "nullable": true,
            "example": null
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-09-19T12:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-09-19T12:00:00.000Z"
          }
        },
        "required": [
          "id",
          "level",
          "name",
          "parentId",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
      "LocationNodeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "3f8c1b1e-3a6f-4f2e-9a1c-6d9f2b7c4e11"
          },
          "level": {
            "type": "string",
            "enum": [
              "COUNTRY",
              "PROVINCE",
              "CITY",
              "NEIGHBORHOOD"
            ],
            "example": "NEIGHBORHOOD"
          },
          "name": {
            "type": "string",
            "example": "Palermo"
          },
          "parentId": {
            "type": "string",
            "nullable": true,
            "example": null
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-09-19T12:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-09-19T12:00:00.000Z"
          },
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LocationNodeDto"
            }
          }
        },
        "required": [
          "id",
          "level",
          "name",
          "parentId",
          "isActive",
          "createdAt",
          "updatedAt",
          "children"
        ]
      },
      "CreateLocationDto": {
        "type": "object",
        "properties": {
          "level": {
            "type": "string",
            "enum": [
              "COUNTRY",
              "PROVINCE",
              "CITY",
              "NEIGHBORHOOD"
            ],
            "example": "NEIGHBORHOOD"
          },
          "name": {
            "type": "string",
            "example": "Palermo",
            "maxLength": 120
          },
          "parentId": {
            "type": "string",
            "nullable": true,
            "description": "Dónde está. Obligatorio salvo para un país (spec: un barrio sin padre es un 400)."
          }
        },
        "required": [
          "level",
          "name",
          "parentId"
        ]
      },
      "RenameLocationDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Palermo Soho",
            "maxLength": 120
          }
        },
        "required": [
          "name"
        ]
      },
...
    }
  }
}